fixed folder structure for API
